Rename tedge_api::messages -> commands
This module contains only command related messages.
