+++ crates/core/c8y_api/src/json_c8y.rs
@@ -6,10 +6,10 @@ use std::collections::HashMap;
 use std::fmt;
 use tedge_api::alarm::ThinEdgeAlarm;
 use tedge_api::alarm::ThinEdgeAlarmData;
+use tedge_api::commands::SoftwareListCommand;
 use tedge_api::entity_store::EntityMetadata;
 use tedge_api::entity_store::EntityType;
 use tedge_api::event::ThinEdgeEvent;
-use tedge_api::messages::SoftwareListCommand;
 use tedge_api::EntityStore;
 use tedge_api::Jsonify;
 use tedge_api::SoftwareModule;
@@ -366,11 +366,11 @@ mod tests {
     use std::collections::HashSet;
     use tedge_api::alarm::ThinEdgeAlarm;
     use tedge_api::alarm::ThinEdgeAlarmData;
+    use tedge_api::commands::SoftwareListCommandPayload;
     use tedge_api::entity_store::EntityExternalId;
     use tedge_api::entity_store::EntityRegistrationMessage;
     use tedge_api::entity_store::InvalidExternalIdError;
     use tedge_api::event::ThinEdgeEventData;
-    use tedge_api::messages::SoftwareListCommandPayload;
     use tedge_api::mqtt_topics::EntityTopicId;
     use tedge_api::mqtt_topics::MqttSchema;
     use test_case::test_case;
