+++ crates/core/plugin_sm/src/plugin_manager.rs
@@ -9,9 +9,9 @@ use std::io::{self};
 use std::path::Path;
 use std::path::PathBuf;
 use std::process::Stdio;
-use tedge_api::messages::CommandStatus;
-use tedge_api::messages::SoftwareListCommand;
-use tedge_api::messages::SoftwareUpdateCommand;
+use tedge_api::commands::CommandStatus;
+use tedge_api::commands::SoftwareListCommand;
+use tedge_api::commands::SoftwareUpdateCommand;
 use tedge_api::SoftwareError;
 use tedge_api::SoftwareType;
 use tedge_api::DEFAULT;
