+++ crates/core/tedge_agent/src/restart_manager/tests.rs
@@ -13,8 +13,8 @@ use tedge_actors::NoConfig;
 use tedge_actors::Sender;
 use tedge_actors::SimpleMessageBox;
 use tedge_actors::SimpleMessageBoxBuilder;
-use tedge_api::messages::CommandStatus;
-use tedge_api::messages::RestartCommandPayload;
+use tedge_api::commands::CommandStatus;
+use tedge_api::commands::RestartCommandPayload;
 use tedge_api::mqtt_topics::EntityTopicId;
 use tedge_api::RestartCommand;
 use tedge_config::SudoCommandBuilder;
