+++ crates/core/tedge_agent/src/software_manager/actor.rs
@@ -23,10 +23,10 @@ use tedge_actors::RuntimeError;
 use tedge_actors::RuntimeRequest;
 use tedge_actors::Sender;
 use tedge_actors::SimpleMessageBox;
-use tedge_api::messages::CommandStatus;
-use tedge_api::messages::SoftwareCommandMetadata;
-use tedge_api::messages::SoftwareListCommand;
-use tedge_api::messages::SoftwareUpdateCommand;
+use tedge_api::commands::CommandStatus;
+use tedge_api::commands::SoftwareCommandMetadata;
+use tedge_api::commands::SoftwareListCommand;
+use tedge_api::commands::SoftwareUpdateCommand;
 use tedge_api::SoftwareType;
 use tedge_config::TEdgeConfigError;
 use tracing::error;
