+++ crates/core/tedge_agent/src/software_manager/tests.rs
@@ -14,14 +14,14 @@ use tedge_actors::NoConfig;
 use tedge_actors::Sender;
 use tedge_actors::SimpleMessageBox;
 use tedge_actors::SimpleMessageBoxBuilder;
-use tedge_api::messages::CommandStatus;
-use tedge_api::messages::SoftwareCommandMetadata;
-use tedge_api::messages::SoftwareListCommand;
-use tedge_api::messages::SoftwareModuleAction;
-use tedge_api::messages::SoftwareModuleItem;
-use tedge_api::messages::SoftwareRequestResponseSoftwareList;
-use tedge_api::messages::SoftwareUpdateCommand;
-use tedge_api::messages::SoftwareUpdateCommandPayload;
+use tedge_api::commands::CommandStatus;
+use tedge_api::commands::SoftwareCommandMetadata;
+use tedge_api::commands::SoftwareListCommand;
+use tedge_api::commands::SoftwareModuleAction;
+use tedge_api::commands::SoftwareModuleItem;
+use tedge_api::commands::SoftwareRequestResponseSoftwareList;
+use tedge_api::commands::SoftwareUpdateCommand;
+use tedge_api::commands::SoftwareUpdateCommandPayload;
 use tedge_api::mqtt_topics::EntityTopicId;
 use tedge_config::SudoCommandBuilder;
 use tedge_config::TEdgeConfigLocation;
