+++ crates/core/tedge_agent/src/tedge_operation_converter/actor.rs
@@ -15,10 +15,10 @@ use tedge_actors::MessageReceiver;
 use tedge_actors::RuntimeError;
 use tedge_actors::Sender;
 use tedge_actors::UnboundedLoggingReceiver;
-use tedge_api::messages::RestartCommand;
-use tedge_api::messages::SoftwareCommandMetadata;
-use tedge_api::messages::SoftwareListCommand;
-use tedge_api::messages::SoftwareUpdateCommand;
+use tedge_api::commands::RestartCommand;
+use tedge_api::commands::SoftwareCommandMetadata;
+use tedge_api::commands::SoftwareListCommand;
+use tedge_api::commands::SoftwareUpdateCommand;
 use tedge_api::mqtt_topics::Channel;
 use tedge_api::mqtt_topics::EntityTopicId;
 use tedge_api::mqtt_topics::MqttSchema;
