+++ crates/core/tedge_agent/src/tedge_operation_converter/tests.rs
@@ -16,14 +16,14 @@ use tedge_actors::RequestEnvelope;
 use tedge_actors::Sender;
 use tedge_actors::SimpleMessageBox;
 use tedge_actors::SimpleMessageBoxBuilder;
-use tedge_api::messages::CommandStatus;
-use tedge_api::messages::RestartCommandPayload;
-use tedge_api::messages::SoftwareCommandMetadata;
-use tedge_api::messages::SoftwareListCommand;
-use tedge_api::messages::SoftwareModuleAction;
-use tedge_api::messages::SoftwareModuleItem;
-use tedge_api::messages::SoftwareRequestResponseSoftwareList;
-use tedge_api::messages::SoftwareUpdateCommandPayload;
+use tedge_api::commands::CommandStatus;
+use tedge_api::commands::RestartCommandPayload;
+use tedge_api::commands::SoftwareCommandMetadata;
+use tedge_api::commands::SoftwareListCommand;
+use tedge_api::commands::SoftwareModuleAction;
+use tedge_api::commands::SoftwareModuleItem;
+use tedge_api::commands::SoftwareRequestResponseSoftwareList;
+use tedge_api::commands::SoftwareUpdateCommandPayload;
 use tedge_api::mqtt_topics::EntityTopicId;
 use tedge_api::mqtt_topics::MqttSchema;
 use tedge_api::workflow::WorkflowSupervisor;
