+++ crates/extensions/c8y_mapper_ext/src/converter.rs
@@ -64,6 +64,11 @@ use std::path::PathBuf;
 use std::sync::Arc;
 use tedge_actors::LoggingSender;
 use tedge_actors::Sender;
+use tedge_api::commands::CommandStatus;
+use tedge_api::commands::RestartCommand;
+use tedge_api::commands::SoftwareCommandMetadata;
+use tedge_api::commands::SoftwareListCommand;
+use tedge_api::commands::SoftwareUpdateCommand;
 use tedge_api::entity_store;
 use tedge_api::entity_store::EntityExternalId;
 use tedge_api::entity_store::EntityRegistrationMessage;
@@ -72,11 +77,6 @@ use tedge_api::entity_store::Error;
 use tedge_api::entity_store::InvalidExternalIdError;
 use tedge_api::event::error::ThinEdgeJsonDeserializerError;
 use tedge_api::event::ThinEdgeEvent;
-use tedge_api::messages::CommandStatus;
-use tedge_api::messages::RestartCommand;
-use tedge_api::messages::SoftwareCommandMetadata;
-use tedge_api::messages::SoftwareListCommand;
-use tedge_api::messages::SoftwareUpdateCommand;
 use tedge_api::mqtt_topics::Channel;
 use tedge_api::mqtt_topics::EntityTopicId;
 use tedge_api::mqtt_topics::IdGenerator;
