+++ crates/extensions/c8y_mapper_ext/src/operations/config_snapshot.rs
@@ -14,8 +14,8 @@ use c8y_http_proxy::messages::CreateEvent;
 use camino::Utf8PathBuf;
 use std::collections::HashMap;
 use tedge_actors::Sender;
-use tedge_api::messages::CommandStatus;
-use tedge_api::messages::ConfigSnapshotCmdPayload;
+use tedge_api::commands::CommandStatus;
+use tedge_api::commands::ConfigSnapshotCmdPayload;
 use tedge_api::mqtt_topics::Channel;
 use tedge_api::mqtt_topics::ChannelFilter;
 use tedge_api::mqtt_topics::EntityFilter;
