+++ crates/extensions/c8y_mapper_ext/src/operations/config_update.rs
@@ -7,10 +7,10 @@ use c8y_api::smartrest::smartrest_serializer::set_operation_executing;
 use c8y_api::smartrest::smartrest_serializer::succeed_operation_no_payload;
 use c8y_api::smartrest::smartrest_serializer::CumulocitySupportedOperations;
 use std::sync::Arc;
+use tedge_api::commands::CommandStatus;
+use tedge_api::commands::ConfigUpdateCmdPayload;
 use tedge_api::entity_store::EntityExternalId;
 use tedge_api::entity_store::EntityMetadata;
-use tedge_api::messages::CommandStatus;
-use tedge_api::messages::ConfigUpdateCmdPayload;
 use tedge_api::mqtt_topics::Channel;
 use tedge_api::mqtt_topics::ChannelFilter;
 use tedge_api::mqtt_topics::EntityFilter;
