+++ crates/extensions/c8y_mapper_ext/src/operations/firmware_update.rs
@@ -6,9 +6,9 @@ use c8y_api::smartrest::smartrest_serializer::fail_operation;
 use c8y_api::smartrest::smartrest_serializer::set_operation_executing;
 use c8y_api::smartrest::smartrest_serializer::succeed_operation_no_payload;
 use c8y_api::smartrest::smartrest_serializer::CumulocitySupportedOperations;
+use tedge_api::commands::FirmwareInfo;
+use tedge_api::commands::FirmwareUpdateCmdPayload;
 use tedge_api::entity_store::EntityExternalId;
-use tedge_api::messages::FirmwareInfo;
-use tedge_api::messages::FirmwareUpdateCmdPayload;
 use tedge_api::mqtt_topics::Channel;
 use tedge_api::mqtt_topics::ChannelFilter::Command;
 use tedge_api::mqtt_topics::ChannelFilter::CommandMetadata;
