+++ crates/extensions/c8y_mapper_ext/src/operations/log_upload.rs
@@ -14,9 +14,9 @@ use c8y_http_proxy::messages::CreateEvent;
 use camino::Utf8PathBuf;
 use std::collections::HashMap;
 use tedge_actors::Sender;
-use tedge_api::messages::CommandStatus;
-use tedge_api::messages::LogMetadata;
-use tedge_api::messages::LogUploadCmdPayload;
+use tedge_api::commands::CommandStatus;
+use tedge_api::commands::LogMetadata;
+use tedge_api::commands::LogUploadCmdPayload;
 use tedge_api::mqtt_topics::Channel;
 use tedge_api::mqtt_topics::ChannelFilter::Command;
 use tedge_api::mqtt_topics::ChannelFilter::CommandMetadata;
