+++ crates/extensions/tedge_config_manager/src/actor.rs
@@ -15,9 +15,9 @@ use tedge_actors::LoggingSender;
 use tedge_actors::MessageReceiver;
 use tedge_actors::RuntimeError;
 use tedge_actors::Sender;
-use tedge_api::messages::CommandStatus;
-use tedge_api::messages::ConfigSnapshotCmdPayload;
-use tedge_api::messages::ConfigUpdateCmdPayload;
+use tedge_api::commands::CommandStatus;
+use tedge_api::commands::ConfigSnapshotCmdPayload;
+use tedge_api::commands::ConfigUpdateCmdPayload;
 use tedge_api::Jsonify;
 use tedge_downloader_ext::DownloadRequest;
 use tedge_downloader_ext::DownloadResult;
