+++ crates/extensions/tedge_log_manager/src/actor.rs
@@ -18,8 +18,8 @@ use tedge_actors::NoMessage;
 use tedge_actors::RuntimeError;
 use tedge_actors::Sender;
 use tedge_actors::SimpleMessageBox;
-use tedge_api::messages::CommandStatus;
-use tedge_api::messages::LogUploadCmdPayload;
+use tedge_api::commands::CommandStatus;
+use tedge_api::commands::LogUploadCmdPayload;
 use tedge_api::Jsonify;
 use tedge_file_system_ext::FsWatchEvent;
 use tedge_mqtt_ext::MqttMessage;
